user friendly

+++ server/src/main/java/com/abc/chenzeshenga/logistics/controller/TrackNoController.java
@@ -87,6 +87,9 @@ import java.util.List;
 
     @GetMapping @RequestMapping("/pk") public Json generate(@RequestParam String channelCode) {
         TrackNo trackNo = trackNoMapper.generate3(channelCode);
+        if(trackNo==null){
+            return Json.fail().msg("获取追踪单号失败，管理员将在后台手动添加追踪单号，您可忽略当前错误");
+        }
         Long minTrackno = Long.valueOf(trackNo.getMin());
         if (minTrackno + 1 > Long.valueOf(trackNo.getMax())) {
             trackNoMapper.delete(trackNo.getId());
